requirements.txt
newsapi-python==0.2.7
pandas==2.0.3
newspaper3k==0.2.8
spacy==3.6.1
transformers==4.30.2
tqdm==4.65.0
colorama==0.4.6
python-dotenv==1.0.0
requests==2.31.0 